requirements.txt
aiogram==3.10.0
gspread==6.1.2
google-api-python-client==2.133.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
certifi==2024.12.14
apscheduler==3.10.4
pymorphy3
python-docx==1.1.2
SQLAlchemy==2.0.49
asyncpg==0.31.0
alembic==1.18.4
